Dev merge release version

+++ Assets/Scripts/GameProcess.cs
@@ -24,6 +24,9 @@ public class GameProcess : MonoBehaviour
 
     public GameObject flagpole;
 
+    public Text timerText;
+    public float currectGameTime;
+
     private void Awake()
     {
         scoreText = GameObject.FindGameObjectWithTag("Score").GetComponent<Text>();
@@ -36,6 +39,13 @@ public class GameProcess : MonoBehaviour
             bossUI.SetActive(false);
         if (stairs != null)
             stairs.SetActive(false);
+
+        currectGameTime = PlayerPrefs.GetFloat("@currentGameTime", 0f);
+        if (SceneManager.GetActiveScene().name == "Tutorial")
+            timerText.gameObject.SetActive(false);
+        else
+            timerText.gameObject.SetActive(true);
+
     }
 
     private void Start()
@@ -65,17 +75,28 @@ public class GameProcess : MonoBehaviour
         }
         enemiesText.text = $"{GameSaving.instance.deadEnemies} / {GameSaving.instance.enemiesCount}";
         scoreText.text = GameSaving.instance.score.ToString();
-        if(PlayerPrefs.GetString("@mode") == "Normal Mode")
+        if(PlayerPrefs.GetString("@mode") == "Normal Mode" && SceneManager.GetActiveScene().name != "Tutorial")
             PlayerPrefs.SetInt("@level", SceneManager.GetActiveScene().buildIndex);
+
+        if (SceneManager.GetActiveScene().name != "Tutorial")
+        {
+            InvokeRepeating("UpdateTimerText", 0f, .5f);
+        }
     }
 
     private void Update()
     {
-        //  if pressed Escape and time not stopped
-        if (Input.GetKeyDown(KeyCode.Escape) && Time.timeScale == 1)
+        //  make pause or unpause
+        if (Input.GetKeyDown(KeyCode.Escape))
         {
             Pause();
         }
+        currectGameTime += Time.deltaTime;
+    }
+
+    private void UpdateTimerText()
+    {
+        timerText.text = GameSaving.instance.ConvertGameTimeToString(currectGameTime);
     }
 
     public void Pause()
@@ -86,14 +107,24 @@ public class GameProcess : MonoBehaviour
 
     public void Restart()
     {
-        DontSaveStatsByMode();
+        ClearStatsByMode();
         pausePanel.SetActive(false);
-        SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
+        if(PlayerPrefs.GetString("@mode") == "Normal Mode")
+        {
+            SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
+            return;
+        }
+
+        currectGameTime = 0f;
+        PlayerPrefs.DeleteKey("@currentGameTime");
+        SceneManager.LoadScene("level 1");
     }
 
     public void Exit()
     {
-        GameSaving.instance.ClearPlayerPrefs();
+        if(PlayerPrefs.GetString("@mode") == "Hard Mode")
+            GameSaving.instance.ClearPlayerPrefs();
+
         pausePanel.SetActive(false);
         Application.Quit();
     }
@@ -127,7 +158,7 @@ public class GameProcess : MonoBehaviour
             GameObject _instance = Instantiate(item, parent.position, Quaternion.identity, parent);
             counter += 1;
         }
-        DontSaveStatsByMode();
+        ClearStatsByMode();
     }
 
     private void OnDestroy()
@@ -141,25 +172,22 @@ public class GameProcess : MonoBehaviour
         GameSaving.instance.OnBossStart -= OnBossStartHandler;
         GameSaving.instance.OnBossDie -= OnBossEndHandler;
         GameSaving.instance.OnEndLevel -= OnEndLevelHandler;
+
+        if(SceneManager.GetActiveScene().name != "Tutorial")
+            PlayerPrefs.SetFloat("@currentGameTime", currectGameTime);
     }
 
     public void GameOverRestart()
     {
-        if (SceneManager.GetActiveScene().buildIndex == 1)
-        {
-            SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
-            return;
-        }
-
         if (PlayerPrefs.GetString("@mode") == "Hard Mode")
         {
-            SceneManager.LoadScene(2);
+            SceneManager.LoadScene("level 1");
             return;
         }
         SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
     }
 
-    private void DontSaveStatsByMode()
+    private void ClearStatsByMode()
     {
         if (PlayerPrefs.GetString("@mode") == "Hard Mode")
         {
@@ -170,10 +198,14 @@ public class GameProcess : MonoBehaviour
 
     public void Menu()
     {
-        PlayerPrefs.DeleteKey("@level");
-        DontSaveStatsByMode();
+        if(gameOverPanel.activeSelf)
+            PlayerPrefs.DeleteKey("@level");
+
+        if(PlayerPrefs.GetString("@mode") == "Hard Mode")
+            ClearStatsByMode();
+
         SoundMusicManager.instance.backgroundMusicStop();
-        SceneManager.LoadScene(0);
+        SceneManager.LoadScene("Menu");
     }
 
     private void OnBossStartHandler()
